fastapi_backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
pydantic>=2.10.4
alembic==1.14.0
python-dotenv==1.0.1
yfinance==0.2.18
reportlab==4.0.4
celery==5.3.4
redis==5.0.1
email-validator
requests
pandas
